outreach-agent/requirements.txt
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
pydantic>=2.6.0
python-dotenv>=1.0.0
